requirements.txt
fastapi
uvicorn
google-generativeai
chromadb
pypdf
python-docx
python-multipart
pydantic
diskcache
loguru
watchdog
aiofiles